+++ SIMS_Projekat_Dijagram_Neaktivnosti/src/kontroleri/PrimerakKontroler.java
@@ -36,7 +36,7 @@ public class PrimerakKontroler {
 		iR.dodajPrimerak(p, izdanje);
 	}
 	
-	public void promeniCenu(int id, int cena) throws IOException {
+	public void promeniCenu(int id, double cena) throws IOException {
 		PrimerakRepo repo = new PrimerakRepo();
 		repo.promeniCenu(id, cena);
 	}
